backend/requirements.txt
# Django Core Dependencies
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework-simplejwt

# Django Dependencies (automatically installed with Django)
asgiref==3.11.0
sqlparse==0.5.3


supabase  
python-dotenv>=1.0.0


#install these if supabase could not be installed properly on windows
# anyio==4.11.0
# asgiref==3.11.0
# certifi==2025.11.12
# deprecation==2.1.0
# Django==5.2.8
# django-cors-headers==4.9.0
# django-filter==25.2
# djangorestframework==3.16.1
# exceptiongroup==1.3.1
# gotrue==2.12.4
# h11==0.16.0
# h2==4.3.0
# hpack==4.1.0
# httpcore==1.0.9
# httpx==0.28.1
# hyperframe==6.1.0
# idna==3.11
# packaging==25.0
# postgrest==1.1.1
# pydantic==1.10.24
# PyJWT==2.10.1
# python-dateutil==2.9.0.post0
# python-dotenv==1.2.1
# realtime==2.5.3
# six==1.17.0
# sniffio==1.3.1
# sqlparse==0.5.3
# storage3==0.12.1
# StrEnum==0.4.15
# supabase==2.16.0
# supafunc==0.10.2
# typing_extensions==4.15.0
# tzdata==2025.2
# websockets==15.0.1